+++ src/actions/index.js
@@ -17,8 +17,6 @@ import * as coverage from "./coverage";
 import * as projectTextSearch from "./project-text-search";
 import * as quickOpen from "./quick-open";
 import * as sourceTree from "./source-tree";
-import * as loadSourceText from "./sources/loadSourceText";
-import * as newSources from "./sources/newSources";
 import * as debuggee from "./debuggee";
 import * as toolbox from "./toolbox";
 
@@ -36,8 +34,6 @@ export default {
   ...projectTextSearch,
   ...quickOpen,
   ...sourceTree,
-  ...loadSourceText,
-  ...newSources,
   ...debuggee,
   ...toolbox
 };
