+++ src/actions/sources/newSources.js
@@ -1,5 +1,64 @@
 // @flow
 
+import { PROMISE } from "../utils/middleware/promise";
+import assert from "../../utils/assert";
+import { remapBreakpoints } from "../breakpoints";
+import { throttle } from "lodash";
+import { setEmptyLines, setOutOfScopeLocations } from "../ast";
+import { syncBreakpoint } from "../breakpoints";
+import { searchSource } from "../project-text-search";
+import { closeActiveSearch } from "../ui";
+
+import { getPrettySourceURL, isLoaded } from "../../utils/source";
+import { createPrettySource } from "../sources/createPrettySource";
+import { loadSourceText } from "./loadSourceText";
+import { selectSource } from "./selectSource";
+
+import { prefs } from "../../utils/prefs";
+import { removeDocument } from "../../utils/editor";
+import {
+  isThirdParty,
+  isMinified,
+  shouldPrettyPrint
+} from "../../utils/source";
+import { getGeneratedLocation } from "../../utils/source-maps";
+import { isOriginalId } from "devtools-source-map";
+import {
+  getSource,
+  getSources,
+  getSourceByURL,
+  getPendingSelectedLocation,
+  getPendingBreakpointsForSource,
+  getSourceTabs,
+  getNewSelectedSourceId,
+  getSelectedLocation,
+  removeSourcesFromTabList,
+  removeSourceFromTabList,
+  getTextSearchQuery,
+  getActiveSearch,
+  getGeneratedSource
+} from "../../selectors";
+
+import type { Source } from "../../types";
+import type { ThunkArgs } from "../types";
+import type { State } from "../../reducers/types";
+
+export type SelectSourceOptions = {
+  tabIndex?: number,
+  location?: { line: number, column?: ?number }
+};
+
+function createSource(originalUrl, generatedSource, sourceMaps): Source {
+  return {
+    url: originalUrl,
+    id: sourceMaps.generatedToOriginalId(generatedSource.id, originalUrl),
+    isPrettyPrinted: false,
+    isWasm: false,
+    isBlackBoxed: false,
+    loadedState: "unloaded"
+  };
+}
+
 // If a request has been made to show this source, go ahead and
 // select it.
 async function checkSelectedSource(state: State, dispatch, source) {
@@ -12,7 +71,8 @@ async function checkSelectedSource(state: State, dispatch, source) {
 
 async function checkPendingBreakpoints(state, dispatch, sourceId) {
   // source may have been modified by selectSource
-  const source = getSource(state, sourceId).toJS();
+  let source = getSource(state, sourceId).toJS();
+
   const pendingBreakpoints = getPendingBreakpointsForSource(state, source.url);
   if (!pendingBreakpoints.size) {
     return;
@@ -32,35 +92,36 @@ async function checkPendingBreakpoints(state, dispatch, sourceId) {
   }
 }
 
-async function loadSourceMap(generatedSource: Source, sourceMaps): Source[] {
+async function loadSourceMap(
+  generatedSource: Source,
+  sourceMaps
+): Promise<Source[]> {
   const urls: Array<any> = await sourceMaps.getOriginalURLs(generatedSource);
 
-  return urls.map(
-    originalUrl =>
-      ({
-        url: originalUrl,
-        id: sourceMaps.generatedToOriginalId(generatedSource.id, originalUrl),
-        isPrettyPrinted: false,
-        isWasm: false,
-        isBlackBoxed: false,
-        loadedState: "unloaded"
-      }: Source)
+  if (!urls) {
+    return [];
+  }
+
+  return urls.map(originalUrl =>
+    createSource(originalUrl, generatedSource, sourceMaps)
   );
 }
 
 async function loadOriginalSources(sources: Source[], sourceMaps) {
-  let originalSources = [];
-
-  for (const source of sources) {
-    originalSources = [
+  return sources.reduce(
+    async (originalSources, source) => [
       ...originalSources,
       ...(await loadSourceMap(source, sourceMaps))
-    ];
-  }
+    ],
+    []
+  );
 }
 
-// adds all of the new sources
-// 3.
+export function newSource(source: Source) {
+  return async ({ dispatch, getState, sourceMaps }: ThunkArgs) => {
+    await dispatch(newSources([source]));
+  };
+}
 
 export function newSources(sources: Source[]) {
   return async ({ dispatch, getState, sourceMaps }: ThunkArgs) => {
@@ -82,7 +143,7 @@ export function newSources(sources: Source[]) {
       sourceMaps
     );
 
-    dispatch({ type: "ADD_SOURCES", sources });
+    dispatch({ type: "ADD_SOURCES", sources: originalSources });
 
     // 4. check for a selected source and start loading it
     for (const source of originalSources) {
@@ -93,6 +154,7 @@ export function newSources(sources: Source[]) {
     // NOTE: it would be nice to make this smarter so that
     // we first show un-adjusted breakpoints and then we adjust the locations.
     const allSources = filteredSources.concat(originalSources);
+
     for (const source of allSources) {
       checkPendingBreakpoints(getState(), dispatch, source.id);
     }
