+++ src/actions/sources/tabs.js
@@ -1,0 +1,110 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
+
+// @flow
+
+/**
+ * Redux actions for the sources state
+ * @module actions/sources
+ */
+
+import { PROMISE } from "../utils/middleware/promise";
+import assert from "../../utils/assert";
+import { remapBreakpoints } from "../breakpoints";
+import { throttle } from "lodash";
+import { setEmptyLines, setOutOfScopeLocations } from "../ast";
+import { syncBreakpoint } from "../breakpoints";
+import { searchSource } from "../project-text-search";
+import { closeActiveSearch } from "../ui";
+
+import { getPrettySourceURL, isLoaded } from "../../utils/source";
+import { createPrettySource } from "./createPrettySource";
+import { loadSourceText } from "./loadSourceText";
+import { selectSource } from "./selectSource";
+
+import { prefs } from "../../utils/prefs";
+import { removeDocument } from "../../utils/editor";
+import {
+  isThirdParty,
+  isMinified,
+  shouldPrettyPrint
+} from "../../utils/source";
+import { getGeneratedLocation } from "../../utils/source-maps";
+import { isOriginalId } from "devtools-source-map";
+import {
+  getSource,
+  getSources,
+  getSourceByURL,
+  getPendingSelectedLocation,
+  getPendingBreakpointsForSource,
+  getSourceTabs,
+  getNewSelectedSourceId,
+  getSelectedLocation,
+  removeSourcesFromTabList,
+  removeSourceFromTabList,
+  getTextSearchQuery,
+  getActiveSearch,
+  getGeneratedSource
+} from "../../selectors";
+
+import type { Source } from "../../types";
+import type { ThunkArgs } from "../types";
+import type { State } from "../../reducers/types";
+
+export type SelectSourceOptions = {
+  tabIndex?: number,
+  location?: { line: number, column?: ?number }
+};
+
+export function addTab(source: Source, tabIndex: number) {
+  return {
+    type: "ADD_TAB",
+    source,
+    tabIndex
+  };
+}
+
+export function moveTab(url: string, tabIndex: number) {
+  return {
+    type: "MOVE_TAB",
+    url,
+    tabIndex
+  };
+}
+
+/**
+ * @memberof actions/sources
+ * @static
+ */
+export function closeTab(url: string) {
+  return ({ dispatch, getState, client }: ThunkArgs) => {
+    removeDocument(url);
+    const tabs = removeSourceFromTabList(getSourceTabs(getState()), url);
+    const sourceId = getNewSelectedSourceId(getState(), tabs);
+
+    dispatch({ type: "CLOSE_TAB", url, tabs });
+    dispatch(selectSource(sourceId));
+  };
+}
+
+/**
+ * @memberof actions/sources
+ * @static
+ */
+export function closeTabs(urls: string[]) {
+  return ({ dispatch, getState, client }: ThunkArgs) => {
+    urls.forEach(url => {
+      const source = getSourceByURL(getState(), url);
+      if (source) {
+        removeDocument(source.get("id"));
+      }
+    });
+
+    const tabs = removeSourcesFromTabList(getSourceTabs(getState()), urls);
+    const sourceId = getNewSelectedSourceId(getState(), tabs);
+
+    dispatch({ type: "CLOSE_TABS", urls, tabs });
+    dispatch(selectSource(sourceId));
+  };
+}
