+++ src/actions/tests/pending-breakpoints.spec.js
@@ -18,29 +18,30 @@ jest.mock("../../utils/prefs", () => ({
   },
   clear: jest.fn()
 }));
-
-jest.mock("../sources/loadSourceText", () => ({
-  loadSourceText: () => ({
-    type: "LOAD_SOURCE_TEXT",
-    source: {
-      id: "id",
-      text: "function foo() {}",
-      contentType: "text/javascript"
-    },
-    value: {
-      id: "id",
-      text: "function () {}",
-      contentType: "text/javascript"
-    }
-  })
-}));
-import "../sources/loadSourceText";
+//
+// jest.mock("../sources/loadSourceText", () => ({
+//   loadSourceText: () => ({
+//     type: "LOAD_SOURCE_TEXT",
+//     source: {
+//       id: "id",
+//       text: "function foo() {}",
+//       contentType: "text/javascript"
+//     },
+//     value: {
+//       id: "id",
+//       text: "function () {}",
+//       contentType: "text/javascript"
+//     }
+//   })
+// }));
+// import "../sources/loadSourceText";
 
 import {
   createStore,
   selectors,
   actions,
-  makeSource
+  makeSource,
+  waitForState
 } from "../../utils/test-head";
 
 import { makePendingLocationId, makeLocationId } from "../../utils/breakpoint";
@@ -229,11 +230,14 @@ describe("initializing with disabled pending breakpoints in prefs", () => {
     };
 
     const expectedId = makeLocationId(expectedLocation);
-    const { getState, dispatch } = createStore(simpleMockThreadClient);
+    const store = createStore(simpleMockThreadClient);
+    const { getState, dispatch } = store;
     const source = makeSource("bar.js");
     await dispatch(actions.newSource(source));
-    const bps = selectors.getBreakpoints(getState());
-    const bp = bps.get(expectedId);
+
+    const bp = await waitForState(store, state =>
+      selectors.getBreakpoints(state).get(expectedId)
+    );
 
     expect(bp.location).toEqual(expectedLocation);
     expect(bp.disabled).toEqual(mockedPendingBreakpoint.disabled);
@@ -249,19 +253,26 @@ describe("adding sources", () => {
   });
 
   it("corresponding breakpoints are added for a single source", async () => {
-    const { dispatch, getState } = createStore(simpleMockThreadClient);
+    const store = createStore(simpleMockThreadClient);
+    const { dispatch, getState } = store;
 
     let bps = selectors.getBreakpoints(getState());
     expect(bps.size).toBe(0);
 
     const source = makeSource("bar.js");
     await dispatch(actions.newSource(source));
+    await waitForState(
+      store,
+      state => selectors.getBreakpoints(state).size == 1
+    );
+
     bps = selectors.getBreakpoints(getState());
     expect(bps.size).toBe(1);
   });
 
   it("add corresponding breakpoints for multiple sources", async () => {
-    const { dispatch, getState } = createStore(simpleMockThreadClient);
+    const store = createStore(simpleMockThreadClient);
+    const { dispatch, getState } = store;
 
     let bps = selectors.getBreakpoints(getState());
     expect(bps.size).toBe(0);
@@ -269,6 +280,12 @@ describe("adding sources", () => {
     const source1 = makeSource("bar.js");
     const source2 = makeSource("foo.js");
     await dispatch(actions.newSources([source1, source2]));
+
+    await waitForState(
+      store,
+      state => selectors.getBreakpoints(state).size == 1
+    );
+
     bps = selectors.getBreakpoints(getState());
     expect(bps.size).toBe(1);
   });
